requirements.txt
# AWS Terraform Chatbot — Python dependencies
# Install with: pip install -r requirements.txt

groq>=0.9.0          # Groq API client (OpenAI-compatible)
python-hcl2>=4.3.2   # Terraform HCL parser (optional but recommended)